Increase debounce delay
This alleviates a race condition whereby card results would be returned
after a user had finished backspacing their search query.

There's a small delay between the first character being backspaced, and
the rest of the text being backspaced. This is a common backspace
behaviour. If the debounce delay is too small, then another fetch
request will be sent off in that gap and then be rendered roughly when
the user has finished backspacing all the text. This then produces a
situation where a new list of cards has come back, even though the query
is now empty - which is a strange disconnect and poor UX.