Fix issue with scrolling by removing h-full
See this stack overflow post for why: https://stackoverflow.com/questions/53566405/javascript-offsetheight-not-updating-from-initial-rendered-height

Essentially, I was seeing too many "reached bottom" triggers happening
from the scroll handler. It was because the offsetHeight was no longer
updating as new pages of results were added, so the calculation always
concluded that we'd reached the bottom as scrollY would always get
bigger.