Emulate real user in fetch options for API request
MTG Singles started blocking the API requests, which broke the functionality of Card Spy. Copying the fetch options that the real website uses into our API request allows us to continue using the MTG Singles API without being blocked.

This works for now, although this approach may need to be revised or extended if MTG Singles improves it's blocking measures in the future.