Add Origin and Referer headers to API call
The previous update to the fetch options enabled successful fetching from localhost, but not from deployed environments. Looking more closely at the request that MTG Singles makes to its own API, it includes Origin and Referer headers that might be used as part of a security check.

This commit adds those headers in the hope that it might fix calling the API on deployed environments.