Log out request headers from hosted deployment
The main suspect in broken calls to the MTG singles is a difference between the headers on localhost/curl and what is on the deployed environment. By updating the API call to go directly against something that will echo back the request headers, I'll be able to compare more accurately between different environments and - hopefully - pinpoint the root cause.